refactor(middleware): rename accessLogger to accessLogWriter
